clip-app/requirements.txt
streamlit
torch
transformers==4.38.0
accelerate
pillow
numpy
pandas
matplotlib
seaborn
requests
urllib3<2.0.0
fastapi
python-multipart
uvicorn
nest_asyncio
certifi
prometheus-client
psutil
kafka-python
websockets
asyncio
redis
redis
uuid